requirements.txt
fastmcp>=0.1.0
pydantic>=2.0.0
pyyaml>=6.0
typing-extensions>=4.8.0
python-dotenv>=1.0.0
openai>=1.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
fastapi>=0.100.0
uvicorn>=0.22.0
sse-starlette>=1.6.0
pydantic-monty>=0.1.0
opensandbox>=0.1.0
opensandbox-server>=0.1.0
httpx>=0.24.0
pytest-asyncio>=0.21.0

